uuids

+++ src/models/listener.rs
@@ -3,11 +3,11 @@ use std::hash::Hash;
 use uuid::Uuid;
 #[derive(Hash, Debug, Serialize, Deserialize)]
 pub struct Listener {
-    pub id: u64,
+    pub id: u128,
 }
 
 impl Listener {
-    pub fn new(id: u64) -> Listener {
+    pub fn new(id: u128) -> Listener {
         Listener { id }
     }
     pub fn new_with_random_uuid() -> Listener {
